Allow to configure replication via bound crontab
Replication cron job was enable by default, which is no longer
compatible with the deployment process of search indexer.

Customizing cron job (to remove it or to change its schedule)
required either to edit crontab in the container attached to the
service `musicbrainz` (which had to be redone each time the
container gets replaced) or to edit `crons.conf` in build files
(conflicting with pulling the latest version of musicbrainz-docker).

This patch solves these two issues by enabling replication only when
the file `compose/replication-cron.yml` is used, and by allowing to
bind `/crons.conf` file (in the container attached to the service
`musicbrainz`) to another file than `default/replication.cron`.

However, enabling cron job this way requires to restart container.