Allow to configure indexer via bound config.ini
Configuring indexer (which is needed to adapt to the CPU and RAM of the
host) required either to edit `config.ini` in the container attached to
the service `indexer` (which had to be redone each time the container
gets replaced) or to edit `config.ini` file under `build/sir
(conflicting with pulling the latest version of musicbrainz-docker).

This patch solves this issue by allowing to bind `/code/config.ini` file
(in the container attached to the service `musicbrainz`) to another file
than `default/indexer.ini`.

It still requires to restart container if live-indexing is on.