Allow to set replication token via Docker secrets
Setting MetaBrainz access token for MusicBrainz slave server
required to edit `DBDefs.pm` either in the container attached to the
service `musicbrainz` (which had to be redone each time the container
got replaced) or in the host at `build/musicbrainz/DBDefs.pm` (being
insecure and conflicting with musicbrainz-docker latest version).

This patch solves this issue by managing the token as Docker secret
which is bound to a file written under `local/secrets/` directory.

It still requires to restart the `musicbrainz` service’s container.