Set PROD_STANDY to read the production database
This is a database definition for use by official maintainers only.

Note that standby/slave/secondary database is read-only anyway, thus
setting user to `musicbrainz` or musicbrainz_ro` makes no difference.
It is set to `musicbrainz_ro` as a precaution just in case the SSH
tunnel mistakenly points to a database other than the standby prod.

It relies on https://github.com/metabrainz/musicbrainz-server/pull/1337