Exit start script on error
This prevents starting the server and cron if no db and redis connections
can be made within 60 seconds or some other error occurs before that.

This also works around a bug in Perl local::lib that has been fixed in 2.000019.
When the base image is updated line 2 can be set straight to `set -eu`.