Install scripts under /usr/local/bin/ instead of /
Running/executing scripts in musicbrainz/search containers was requiring
to prefix with a slash, which is both unnecessary and barely portable.

This patch installs scripts that are expected to be run from the
command-line (in `musicbrainz` and `search`) under `/usr/loca/bin/`
which is included in `PATH` environment variable by default.

Instructions, scripts, and cron task have been updated accordingly.

Upgrade notes:

If you customized replication schedule,
update it to run `replication.sh` without the leading slash.