Try reaching different PGP servers/pools if needed
Previously, refreshing PGP keys was retried with the same server pool.
It may happen that this server (actually pool) could not be reached,
either due to availability issues or due to local networking issues.

This patch exposes PGP_SERVERS build argument that is a space-delimited
list of servers (or server pools) to be reached to refresh PGP keys.
It attempts to refresh keys up to 3 times per server at most.

It also prevents gpg to send innocuous messages to the error output,
and shows a more helpful error message in case of failure.