Set Postgres configuration parameters from command
The two additional configuration parameters were hardcoded in a script
installed in the Docker image for the service 'db'. It was not easy to
change 'shared_buffers' value to meet meet resources and satisfy needs.

This patch drops the script and set configuration parameters from the
'command' defined for the db service. It allows to change these values
without having to rebuild images, and to add any other configuration
parameter or option: just override 'command' in a local compose file.

It also increases the value for shared_buffers to 2048MB by default,
which is more suitable when running a mirror to speed web service up.

References:
* https://hub.docker.com/_/postgres/
* https://www.postgresql.org/docs/9.5/app-postgres.html