Add env variable and override for remote search
Add the environment variable `MUSICBRAINZ_SEARCH_SERVER` to the image
for the service `musicbrainz`.  That can be set to a remote service
such as `search.musicbrainz.org` through a Docker compose override.

    musicbrainz:
      environment:
        MUSICBRAINZ_SEARCH_SERVER=search.musicbrainz.org

The services `indexer`, `mq`, and `search` are still required for now,
because these are referred to in docker-compose.yml and some overrides.