Set SHM size to 2 GB for database container
Default SHM size in Docker is 64MB while new Postgres 10+ uses /dev/shm
in order to handle additional parallelism in query execution. This makes
the database to throw error "could not resize shared memory segment".

This bumps the allowed SHM size for the database up to 2 GB which is
more than enough for a mirror and provisions to handle large queries.