MBVM-69: Fix persistence of search message queues
Message queues for live indexing were not found after recreating
container for `mq` service. This caused live indexing to stop working.
A known workaround was to run `amqp_setup` again.
This was due to RabbitMQ storing its data under a directory named after
the host name which is sets at random on container’s creation.

This patch sets hostname to `mq` in new containers for `mq` service.

Deploying this patch requires special steps only if live indexing is on:

1. Stop `musicbrainz` service to stop replication
    sudo docker-compose stop musicbrainz

   (Or stop `cron` and `replication.sh` if running)

2. Wait for `search.index` queue to become empty
    sudo docker-compose exec mq \
        rabbitmqadmin -u sir -p sir -V /search-index-rebuilder \
            list queues

3. Recreate container for `mq` service to use the new hostname
    sudo docker-compose up -d mq

4. Recreate `search` queues
    sudo docker-compose exec indexer python -m sir amqp_setup

5. Check that `search.index` queue is available (and empty)
    sudo docker-compose exec mq \
        rabbitmqadmin -u sir -p sir -V /search-index-rebuilder \
            list queues

6. Restart the indexer
    sudo docker-compose restart indexer

7. Start `musicbrainz` service
    sudo docker-compose start musicbrainz

   (Or resume replication)