Secure published ports in development setup
By default published ports are listening on all interfaces of the host,
making those accessible from the outside in the absence of a firewall.
Even though it is the desired behavior when deploying a mirror with
replication, it might not be wanted for development or other use case.

This patch allows specifying the host interface to listen to, and
recommend listening to the host loopback only for development.

Ref: https://github.com/metabrainz/guidelines/blob/master/Docker.md