MBVM-84: Database-only mirror alternative (#226)
The alternative base compose file is placed into the same directory as
the default base compose file so as not to break relative paths set in
override files such as `compose/replication-token.yml`.

The script `admin/compose` is updated to handle this alternative too.

Note that a Redis instance is currently still required even though it is
not actually used. See the ticket MBS-12440 for follow-up.

Upgrade notes: If you want to switch a running mirror to this base
compose file, you will have to re-add other compose files you had.
Check with `admin/configure show` before and after running the command.