Replace virtualenv with user site dir for SIR dev (#230)
For SIR development setup only, so far its dependencies were installed
in a Python virtual environment, without enabling it (to allow running
command with `docker-compose exec`) which was a hack that is no longer
working with the new Python base image 2.7-20220421 used since SIR 3.

Instead this patch sets user Python site-packages directory to point to
the same location `/code/venv-musicbrainz-docker` (which is matched by
`sir/.gitignore`) and install the dependencies to it. This change should
be transparent to previously installed SIR development setup.