Fix fatal error with renderer in development setup
When starting the renderer for MusicBrainz Server in development setup,
it runs some `git` commands to gather information for display. However,
a fatal error occurs with recent versions of `git` (currently 2.25.1)
due to the directory `/musicbrainz-server` being owned by a different
user as it is mounted from the host filesystem.

This patch sets git config in the image to prevent this error.

To update, if and only if you are using a developement setup, you have
to replace the image and the container for the `musicbrainz` service.