MBVM-89: Show info about signing up on first fetch
It shows information about signing up when fetching data (database
and/or search indexes) dumps, for commercial or non-commercial use.

It doesn’t show it for fetching sample dump which is for development.

It doesn’t show it again for further downloads if answered already.
To do that, it creates an empty file in the database dump directory.
If we need to update any of these messages in the future,
we can compare the timestamp of this file to show the updated message.