Amend 9ad01271: Use Wget option -c instead of -N
The Wget option `-N` (that enables time-stamping) skips already fully
retrieved file, but doesn't resume downloading partially retrieved file.

This patch replaces it with the option `-c` that does that. However,
this option assumes that any local file is a previous download attempt
of the same remote file by the same name. This is not true if
downloading a different version of the dump. To prevent this, this patch
also cleans up the download directory beforehand if the version differs.