Explicitely set ulimits for mq to ERL_MAX_PORTS
On Linux ERL_MAX_PORTS defaults to 65,536. But the Docker container for
the service `mq` rather uses the host ulimits. When ulimits is infinite,
it can make the container to take a very long time to start while using
a lot of CPU for nothing. It depends on the host settings.

This patch explicitely sets ulimits to 65,536 as recommended for `mq`.

See https://github.com/docker-library/rabbitmq/issues/545 for details.