Update disk requirements for search indexes
On the suggestion of @PeterCodar, and after @reosarevok double-checked
that setting up a mirror with pre-built search indexes takes 215 GB,
the recommended disk space is bumped to 250 GB to provision for growth.

Note that pre-built search index archives take 45 GB, so you don’t need
that much if your host is powerful enough to build search indexes.