MBVM-42 (1/2): Prebuild musicbrainz service image
The last service to still be fully built from source was musicbrainz.
It was taking some time (as Perl modules were installed sequentially,
and now even Perl itself is compiled from source), and happened to break
after release time on unexpected upstream changes.

This patch makes deploying mirror to use the prebuilt image
`metabrainz/musicbrainz-docker-musicbrainz:<tag>` for the service
musicbrainz by default instead of building it from the source.

This image has to be built and pushed by the repository maintainers,
using the newly added script, before each release. It is always trying
to pull the latest upstream Docker image for security updates.

For backward compatibility, the image tag isn’t set in the compose
files directly, it is set in a Dockerfile instead. It makes it
incidentally easier to customize the image for particular needs.
It will be revisited when merging into musicbrainz-server repository.