MBVM-42 (2/2): Prebuild db service image
The pg_amqp extension was still built from source for the db service.
It was taking some time for compilation, and happened to break after
release time on unexpected upstream changes.

This patch makes deploying mirror to use the prebuilt image
`metabrainz/musicbrainz-docker-db:<tag>` for the service
db by default instead of building pg_amqp from the source.

This image has to be built and pushed by the repository maintainers,
using the newly added script, before each release. It is always trying
to pull the latest upstream Docker image for security updates.

For backward compatibility, the image tag isn’t set in the compose
files directly, it is set in a Dockerfile instead. It makes it
incidentally easier to customize the image for particular needs.
It will be revisited when merging into musicbrainz-server repository.