Only run table counts every 5 minutes
There seems to be no good reason why we would keep hitting the DB
every 30 seconds to get the counts. 5 minutes seems more than enough.

My understanding is that if I set min_interval here to 300s (5m)
it will just keep the value for that long and keep responding with it,
however often prometheus asks.