Allow running any version of SIR via configuration
Previously, running a different version of SIR was requiring either to
modify the Dockerfile or to add a local compose file setting build arguments.

This patch allows specifying the version of SIR, the Python version and
the date of its base image, through build arguments and Docker Compose
environment variables for SIR development setup only as needed:

- `SIR_DEV_PYTHON_VERSION` (matching a `metabrainz/python` image tag)
- `SIR_DEV_BASE_IMAGE_DATE` (matching a `metabrainz/python` image tag)
- `SIR_DEV_VERSION` (informative only)

Rebuilding the image is required after changing build arguments.