[WIP] Update sir/solr to py3/solr9 versions
This requires command start-musicbrainz-solrcloud - we can add
it to the docker-compose.yml or to the Dockerfile.

This removes v from the the sir branch / version - we just add it as
part of the version name for MBS and I don't see why we force
everything to start with v in Sir. I needed to remove it temporarily
at least so I could test my branch
that does not start with v, but it seems sensible long-term IMO.

WIP because the solr and sir versions should be replaced
with the final ones.