Amend 5b1cef198: Fix loading backup archives again
When loading (SolrCloud) backup archives again, Solr does remove the
previously restored content of each collection excepted the collection
`recording` which happens to have more than one shard and more than 50G.

This behavior is currently neither documented nor reported as a bug.
It has to be reproduced and reported with the latest version of Solr.

In the meantime, this patch provides a workaround which consists in
removing the volume containing Solr data beforehand. That should prevent
filling up the device storing Docker volumes as in the issue #316.

Reference: https://solr.apache.org/guide/solr/9_7/deployment-guide/collection-management.html#restore