Add "Connection errors during reindex" to TROUBLESHOOTING.md
Add a new section, "Connection errors during reindex". It describes the problems I had getting the reindex operation to run successfully, and the changes to memory limits which worked around the problem. 

It gives examples of the error messages arising from the problems, in the hopes of providing search engine fodder for someone else experiencing the same problem.